backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Supabase
supabase==2.10.0
httpx==0.27.2

# Auth & Security
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-multipart
